Conflicts:
	StudentRanking/StudentRanking.v12.suo
	StudentRanking/StudentRanking/App_Data/aspnet-StudentRanking-20150510184220.mdf
	StudentRanking/StudentRanking/App_Data/aspnet-StudentRanking-20150510184220_log.ldf
	StudentRanking/StudentRanking/Web.config
	StudentRanking/StudentRanking/bin/StudentRanking.dll
	StudentRanking/StudentRanking/bin/StudentRanking.dll.config
	StudentRanking/StudentRanking/bin/StudentRanking.pdb
	StudentRanking/StudentRanking/obj/Debug/StudentRanking.csproj.FileListAbsolute.txt
	StudentRanking/StudentRanking/obj/Debug/StudentRanking.csproj.GenerateResource.Cache
	StudentRanking/StudentRanking/obj/Debug/StudentRanking.csprojResolveAssemblyReference.cache
	StudentRanking/StudentRanking/obj/Debug/StudentRanking.dll

+++ StudentRanking/StudentRanking/Models/ProgrammeRules.cs
@@ -13,9 +13,9 @@ namespace StudentRanking.Models
         public String ProgrammeName{ get; set; }
 
         [Required]
-        public ushort MaleCount{ get; set; }
+        public int MaleCount{ get; set; }
 
         [Required]
-        public ushort FemaleCount{ get; set; }
+        public int FemaleCount{ get; set; }
     }
 }
